+++ lib/in_date_utils.dart
@@ -1,3 +1,3 @@
 library in_date_utils;
 
-export 'src/date_utils.dart';
+export 'src/date_time_utils.dart';